+++ src/unidl/__init__.py
@@ -1,3 +1,3 @@
 """UniDL - an integrated streaming service browser and download application."""
 
-__version__ = "2.1.3"
+__version__ = "2.1.4"